Prevent Live Caption from captioning muted videos that were not manually played
This CL prevents the Live Caption feature from captioning muted videos
that were not manually played. Previously, the feature used
AutoplayPolicy::WasAutoplayInitiated to determine whether the playback
was initiated by autoplay. AutoplayPolicy::WasAutoplayInitiated is used
to record metrics and only considers a playback as initiated by autoplay
if it was explicitly allowed by the autoplay policy because the website
has a high Media Engagement Index (MEI) score. Videos that autoplay
because the website received a user gesture are not considered as
initiated by autoplay.

For the purposes of Live Caption, a video should be considered manually
played only if the video was played with a transient user activation.

Bug: 1280059
Change-Id: Ic67ad93b023c554328496a83533118f08dd59ca7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3340085
Reviewed-by: Kenneth MacKay <kmackay@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Evan Liu <evliu@google.com>
Cr-Commit-Position: refs/heads/main@{#955335}
NOKEYCHECK=True
GitOrigin-RevId: 4a95d63ab51aa5090b1ee0c18efa4a154d6f110c