Initialize non-existing frame with outputted set to true
When there is a gap in frame nums, the D3D11Video decoder crashes. This issue is caused by the DPB being full because frames are not properly deleted.
The h264 spec (08/2021 C2.1) states:

"After the marking of each generated frame, each picture m marked by the "sliding window" process as "unused for
reference" is removed from the DPB when it is also marked as "non-existing" or its DPB output time is less than or equal
to the CPB removal time of the current picture n; i.e., to,dpb( m ) <= tr( n ), or it has OutputFlag equal to 0. When a frame or
the last field in a frame buffer is removed from the DPB, the DPB fullness is decremented by one. The "non-existing"
generated frame is inserted into the DPB and the DPB fullness is incremented by one."

Currently, only outputted frames are deleted from the DPB while non-existing frames (generated to fill frame_num gap) are not marked as outputted.
The chosen approach is the same as the H264 reference decoder, namely setting the outputted property on non-existing frames to true. (CF. ldecod/src/mbuffer.c 2739)

An alternative solution would be to extend the check for deletion, allowing non-existing, non-outputted frames to be deleted as well.

Bug: 1275588
Change-Id: Ib48f5135b046fbc3fe4c614a4b1e262a8d712d17
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3339012
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Dan Sanders <sandersd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#957702}
NOKEYCHECK=True
GitOrigin-RevId: d9400bfab346032dcd3baccaf1e88a9b4e0e002b