Fix issue with VideoFrameMetadata merging for bool=true fields.
The code was written assuming that if (metadata.value) was false,
it shouldn't be merged. This is fine for optional fields, but not
bool fields.

Specifically in this bug, `texture_origin_is_top_left` defaults to
true, but is set to false when a canvas backed VideoFrame is made.
The old path would fail to copy the false value over.

This fixes the merge process and adds a test for this case. It also
adds a missing 'dcomp_surface' to the merge process.

R=tguilbert

Fixed: 1277865
Change-Id: I96f7a8032327adaa150ffe7412883bcfb051f71e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3385357
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Auto-Submit: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Cr-Commit-Position: refs/heads/main@{#958843}
NOKEYCHECK=True
GitOrigin-RevId: adf93dc59f84d4efb3c060585cc80b0542e46e45