Reland "media/gpu/vaapiVED: Remove VASurface in EncodeResult"
This reland CL fixes a bug that accessing absl::nullopt in
VaapiWrapper::DownloadFromVABuffer.

============Original commit description============

EncodeResult owns VASurface to be synced in DownloadFromVABuffer().
The surface has been synced in GetEncodeChunkSize(). So the sync
call in DownloadFromVABuffer() is not necessary. We remove VASurface
in EncodeResult.

Bug: b:207088069
Test: video_encode_accelerator_tests on volteer and grunt
Change-Id: I4cbd1d9109d2e176600bd32b0338b88d0ee8d575
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3367560
Reviewed-by: Miguel Casas-Sanchez <mcasas@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#959026}
NOKEYCHECK=True
GitOrigin-RevId: 7384d7c6036638362c574e21908ea29f6cda9d59