Reland "media/gpu/vaapiVED: Replace VASurface in EncodeJob with its id and size"
EncodeJob has to own a VASurface that has encoder input so that the
encode process for EncodeJob can be executed any time. Now it is
in VaapiVideoEncoderDelegate::Encode() that the encode for EncodeJob
is complete. If the input VASurface is valid when EncodeJob is
complete, we don't need to keep VASurface in EncodeJob and would
rather have VASurfaceID and size of the VASurface, which are used
in Encode().

Bug: b:207088069
Test: video_encode_accelerator_tests
Change-Id: If83311a6993c0a2793df519c93156e51fe2568d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3369044
Reviewed-by: Miguel Casas-Sanchez <mcasas@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#959034}
NOKEYCHECK=True
GitOrigin-RevId: 110f3e2c1ed7ab6789bf452613ccb6b8ae328a15