media/gpu/vaapi: render node selection for media usecases
Depending on what driver gets loaded first, renderD128 and renderD129
are chosen between i915 and vgem. However, we are making an assumption
here that renderD128 will always be i915 driver and this is incorrect.
Changes here ensures that the first render node that is not vgem will be
chosen for media workloads.

BUG=b:216220225
TEST=iHD media driver should be used during VP9 video playback

Signed-off-by: Sushma Venkatesh Reddy <sushma.venkatesh.reddy@intel.com>
Change-Id: Id5eb9ed3a5e7eed5e9cf326c35a4e6dc4ec52850
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3417188
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#963883}
NOKEYCHECK=True
GitOrigin-RevId: 1569a7474efe6b340322d7e97ae4534b60fb835c