Allow only one cause for a status.
This is required in order to make it possible to get a "root cause"
for a status that makes sense. If a status is allowed to have multiple
causes, it becomes hard to tell what the real root failure is.

In the theoretical case of "we tried 10 decoders and all failed", I
think the best thing to do would be to add the failures as data, keyed
perhaps on decoder name?

R=liberato

Change-Id: I50f64eb5e600bb646210d863466577429536f727
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3428901
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Ted (Chromium) Meyer <tmathmeyer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967514}
NOKEYCHECK=True
GitOrigin-RevId: 791e739cd36c8483da32ad49a052296a0b872217