MSE: Fail if timestampOffset makes PTS or DTS infinite
Even if timestampOffset, PTS and DTS are all finite, the application of
the offset to PTS or DTS may saturate them to -Infinity or +Infinity.
This change fails frame processing if an infinite PTS or DTS occurs as a
result of timestampOffset application.

This change fixes the precise case in crbug.com/1178180. It is also
related to the fixes for crbug.com/908222, which is not yet fixed even
with this change.

BUG=1178180,908222

Change-Id: If94a0908974e661289c2162eca2ac34b853471f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3445455
Reviewed-by: Will Cassella <cassew@chromium.org>
Commit-Queue: Matthew Wolenetz <wolenetz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#969174}
NOKEYCHECK=True
GitOrigin-RevId: ac37fda26ea618e5c1d531da29426a27d8c37941