media/gpu, ChromeOS: restate YV12 to kPreferredRenderableFourccs
crrev/c/3390191 removed YV12 from the list of renderable Fourccs,
and that broke ChromeStackDecoding on hana (MTK 8173). This was an
oversight: hana ImageProcessor doesn't support NV12 output, only
YV12. The breakage is not a big deal since the new VD hasn't been
launched on Hana, but is good to have ChromeStackDecoding working
as much as possible.

Essentially this CL reverts then mentioned change, and takes the
chance to add/improve some logging that I fixed on the course of
the investigation.

Bug: b:214601194
Change-Id: Ib4824d6235f2c6ccfb05b5069d942933a4f122e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3457348
Reviewed-by: Steve Cho <stevecho@chromium.org>
Commit-Queue: Miguel Casas-Sanchez <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#970255}
NOKEYCHECK=True
GitOrigin-RevId: ab2fd2048cd26634aaa5c6d8465c3bda05099eea