media/gpu/vaapi_unittest: DRM primary node selection
Depending on what driver gets loaded first, card0 and card1 are chosen
between i915 and vgem. However, we are making an assumption here that
card0 will always be i915 driver and this is incorrect. Changes here
ensures that the first primary node that is not vgem will be chosen for
media testcases.

BUG=b:219727232
TEST=graphics.VAAPIUnittest.common test should pass

Signed-off-by: Sushma Venkatesh Reddy <sushma.venkatesh.reddy@intel.com>
Change-Id: I1bfc6c085e62ea42cab65378132704ca4d547fe8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3466094
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#971986}
NOKEYCHECK=True
GitOrigin-RevId: d3bc0df4b2e0cf3d6330c7616bd36456cca0fc08