Add ability to control DSP based audio effects in CrAS via field-trials
This CL adds the control logic in Chrome to provide finch-control over
the AEC, NS and AGC, DSP effects in CrAS, allowing safe fallbacks to the
Browser counterparts.

The functionality is such that when the effects are handled by CrAS,
Chrome signals whether CrAS is allowed to offload those to the DSP.
However, even though Chrome states it is ok to offload, it is still
up to CrAS to choose whether to actually offload or not.

The CrAS CL that adds functionality in libcras for receiving the information
passed regarding this is
https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3338425

Bug: b/191697965
Change-Id: I548007ecd5ffc7384532d805d3bf5e55f7f36153
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3330293
Reviewed-by: Henrik Andreasson <henrika@chromium.org>
Commit-Queue: Per Åhgren <peah@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972315}
NOKEYCHECK=True
GitOrigin-RevId: eb299dc225e468fad7144b845534993464f7ab89