[Surface Capture] Fix overlay rendering
Currently, when a video capture overlay is changed it causes a refresh
request--which is not intended to quickly produce a new frame and
can take up to a few seconds for this request to be fulfilled. In
practice, this means that overlays such as the mouse cursor overlay
result in choppy looking video--the cursor "jumps" around at
approximately one frame per second.

This is due to two major factors: a bug time scale issue in the
VideoCaptureOracle setting the maximum refresh window to 1,000 seconds
instead of 1 second, and the refresh logic being inappropriate for this
use case due to the use of a refresh timer to throttle refresh requests.

This patch cleans up the oracle and gets rid of risky constants that are
stored as primitives in a variety of timescales (seconds, milliseconds,
and microseconds) in favor of storing them as base::TimeDelta parameters
with the correct units in instantiation.

This patch also introduces the concept of refresh demands, which unlike
requests that are filled in a very casual and relaxed manner, are
fulfilled immediately as long as a frame isn't already being drawn.
The video capture overlay now demands a refresh instead of queueing one,
meaning that it can capture as frequently as the oracle's minimum
capture period, making smooth 30FPS capture possible.

Bug: 1299140
Change-Id: I3cf571551b61df37f63e6610ce1a71f52c458e95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3481738
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974478}
NOKEYCHECK=True
GitOrigin-RevId: 7fb5c6a516edee52173b681f0537762e7d233611