media/VP9Decoder/Parser: support context-less parsing
The VP9Decoder/Parser couple work either not parsing uncompressed
header at all, or parsing it for the driver and then requesting it
for updated probability tables (commonly referred to as context).
The V4L2 stable API has a third more in which the driver does not
provide those tables back (the control for it is removed), so the
Parser needs to parse those itself and keep them around.

This CL adds support for that to VP9Decoder/Parser via a new ctor
and member variable. This new code should not change any current
functionality: VP9Decoder/Parser are used on D3D11 (win), VA-API
(Intel/AMD) and V4L2 platforms (legacy and "chromium"), but they
should stay the same [1]. The next CL will use the new combination.

This CL is a cleanup of crrev/c/3448742 and crrev/c/3502896, where
nicolas.dufresne@collabora.com and briannorris@chromium.org did
all the heavy lifting. All glory to them and blame for breakages
to me :)

[1] https://source.chromium.org/search?q=file:media%2Fgpu%20IsFrameContextRequired%20file:%5C.cc&sq=

Bug: b:214332361
Change-Id: Ia9cbea88c20429e47269f7ea251adf9a3c082074
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3504034
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Miguel Casas-Sanchez <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977934}
NOKEYCHECK=True
GitOrigin-RevId: 71b9dc43958a57a8f068119dd9a0988bdc95b29a