media/gpu/v4l2: Introduce V4L2VideoDecoderDelegateVP9
This CL introduces a new V4L2 delegate for the VP9 Stable API. This
API was introduced in v5.17 and is being backported in the context of
rk3399 kernel uprev to v.5.10. Using the new delegate or not depends
on a control (V4L2_CID_STATELESS_VP9_FRAME) detection.

This new delegate uses the functionality landed in the previous CL
to instruct VP9Parser to not query the driver for context updates.
Also, it's heavily modelled after the sibling Chromium/Legacy delegates.

This CL is a cleanup of crrev/c/3448742 and crrev/c/3502896, where
nicolas.dufresne@collabora.com and briannorris@chromium.org did
all the heavy lifting. All glory to them and blame for breakages
to me :)

How to test it? On a kevin-kernelnext w/ crrev/c/3410447 patched in,
it should md5sum-verify the otherwise known-broken video sequence
in Tast video.ChromeStackDecoding.vp9_files_from_bugs, but also the
bug can repro by just navigating to youtube.com/watch?v=coYw-eVU0Ks
and/or using the crosvideo.appspot.com VP9 sequences.

Bug: b:214332361
Change-Id: Ie7f0f2b2a127f44bc447a68abec3d94382eec289
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3500195
Reviewed-by: Steve Cho <stevecho@chromium.org>
Commit-Queue: Miguel Casas-Sanchez <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978313}
NOKEYCHECK=True
GitOrigin-RevId: 01e2514e730f0137be5f4ccc0699a11994b42f1e