Add MF AAC Encoder.
This CL introduces MFAudioEncoder which uses the Windows Media
Foundation's AAC encoder.
https://docs.microsoft.com/en-us/windows/win32/medfound/aac-encoder

The MF AAC encoder was designed to encode files, where all the data
is available at the start, and so the API is not perfectly suited
for stream encoding. Specifically, there is no callback or event
to let us know when output data is ready, or when it is ready for more
input.

We make this work with a state machine that is described by this
diagram:
https://mermaid.live/view/#pako:eNqFVNtu2kAQ_ZXpPkRGCn7gkbZIaXGkqJWSBt66lbXYA1jYa3cvrZDNv3cv1Dcc5Ynd2Tkz55wZXJOkTJEsyUGw6gjbNeUA3_GAPP1JiT8sP-3Eal0qhSnkGUcJTCC8lFJtmTxJSn5Z0FeW5yhgPl_BY67lcRSLuG300UbdM8zD-aqRiimEz3BysYwfGtiK84soE5TyWatKK4vwYAcx598aNT7x9q27206Nu8ZhZQo2sLEN6pqSbA_XZqbbU5ojJZeLxbsMB9yzXGIDr6i04EFASQe4MjIEwZoBwV90JrBcIEvPIDTn5nFGyWzWK2olKqGxr6olPgo5Cq0fXcMGHpIEK2WOdf3hKg6LSp3h7s6zQWePiMM201X0-tpYX2PES304ftH7PQpM61qyospRxhmPr8VgBUXGp0p4QWM1N1I28fO3HvfJnEiIUhg29icIWvHu7o0cEp12tL8nw-xOsJ14r4O9-gbjMm5Z_fHhD8tytsvN_rQOj168P6Ngv-8PO6_ITquu-7N7Gziwt5N2Q3Jg3mLavVuQnYpfmtIF42QXh6-aN2MuFt6R7zEberx4b3f6NcLOl4n_w6jweHqLyfHdgN5eEHJPChQFy1LzuastmBJ1xMJ8CZbmmDJxooTyi8nTVWo6RWmmSkGWttw9YVqVmzNP_t99zjpj5stZ-ODlH1c0454

Basically the process is as follows:
1. If we have input in our queue, give it to the encoder until we run
   out or the encoder rejects it (because it is full)
2. Check the encoder for output until it stops producing
3. If, by our count, the encoder could produce more data, check again
4. If we have more input in the queue, by now, go back to 1

Each transition (e.g. from step 1 to step 2) is accomplished by
posting a task to a SequencedTaskRunner. This ensures that we can
safely access our mutable members (like the queue) without issue,
and we yield the thread frequently so we don't block for extended
periods of time.

This class will be used by MojoAudioEncoder to enable encoding to be
done in the GPU process when using WebCodecs and other consumers. See
these other related CLs:
3243816: webcodecs: Mojo interface, service and client for AudioEncoder | https://chromium-review.googlesource.com/c/chromium/src/+/3243816
3308655: webcodecs: Binding for Mojo Audio Encoder | https://chromium-review.googlesource.com/c/chromium/src/+/3308655

Bug: 1072056
Change-Id: I89b6a21407e5d7ff24a09569b34ffbdd8073ad67
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3293969
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Eugene Zemtsov <eugene@chromium.org>
Commit-Queue: Eugene Zemtsov <eugene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978595}
NOKEYCHECK=True
GitOrigin-RevId: cbe1567c04cbac7cc8ebdd6b9d05603599f473fa