media/gpu: implement OOPVideoDecoder
This CL introduces the OOPVideoDecoder, based on the
media::VideoDecoderMixin interface. This decoder acts as a proxy between
the clients and an out-of-process video decoder. It will be used by
Lacros and by Out-of-process Video Decode acceleration.

It also implements the necessary Mojo plumbing to communicate
the client and the decoder process.

Unit tests will be added in a follow-up CL.

This CL does not change ToT behaviour.

Bug: b:171813538
Change-Id: I39baacd2984ffaaff2fffc1dd620e4d257905bb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3094628
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Miguel Casas-Sanchez <mcasas@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#982956}
NOKEYCHECK=True
GitOrigin-RevId: 87fd251713a04532ad4d2a900c2ed466b9ddceed