media::AudioProcessor Remove too strict capture/playout sequence checks
When audio processing runs in the audio service, these function invocations happen on a platform thread not controlled by Chrome. There are no guarantees that the thread does not change on, for example, device changes.
This is a possible cause of https://crbug.com/1306022 .

Bug: chromium:1306022
Change-Id: I8fb6ead008cda7eee0ce02bb263eefa4f1177bca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3532073
Reviewed-by: Olga Sharonova <olka@chromium.org>
Commit-Queue: Sam Zackrisson <saza@chromium.org>
Cr-Commit-Position: refs/heads/main@{#983235}
NOKEYCHECK=True
GitOrigin-RevId: 5d9a67f7e21a7db1561a02154bab201855ff476d