Don't use cpp_proxy_target for stable_video_decoder.
This CL removes the cpp_proxy_target option from the
stable_video_decoder mojom target. It's not clear to me why
the stable_video_decoder target would need to be part of the
//media/mojo/services component, and doing so would cause a dependency
cycle in the follow-up CL:3138304 because in that CL we'll need the
following dependency:

  //media/mojo/mojom:mojom ->
  //media/mojo/mojom/stable:stable_video_decoder

By using cpp_proxy_target = "//media/mojo/services", we're introducing
the following dependency:

  //media/mojo/mojom/stable:stable_video_decoder ->
  //media/mojo/services:services

The dependency chain continues because of existing code:

  //media/mojo/services:services ->
  //chromeos/components/cdm_factory_daemon:cdm_factory_daemon_gpu ->
  //media/mojo/mojom:mojom

And thus, we'd end up with a cycle. This CL breaks that cycle by getting
rid of the dependency of //media/mojo/mojom/stable:stable_video_decoder
on //media/mojo/services:services via the removal of the
cpp_proxy_target.

Bug: b:195769334
Test: None
Change-Id: I053561546350ae1b8fff26fc4d668f7240817b80
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3537159
Reviewed-by: Pilar Molina Lopez <pmolinalopez@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#983647}
NOKEYCHECK=True
GitOrigin-RevId: 1cc702498675512f365aebd63365c742a602adae