[Cast Streaming] change default playout delay
This patch changes the default playout delay for cast streaming
from 100ms to 400ms--this value is usually set to 400 anyway but
the default should be changed as a tech debt cleanup.

This patch also cleans up some primitive types usage of time
types. We have seen several bugs in the past from mismatching
milliseconds and microseconds, so this changes primitives that
use naming for time type (e.g. FooBarMs) to instead be
base::TimeDelta with explicit casts using ToMilliseconds() when a
primitive type is needed. This also has the added benefit of
making refactoring to a different basis, e.g. from
base::Milliseconds(1) to base::Microseconds(80) trivial instead
of a significant refactor.

Bug: 1304761
Change-Id: Ibede50c109787827dc8698ea07083fa8e5cd083b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3532849
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Cr-Commit-Position: refs/heads/main@{#984483}
NOKEYCHECK=True
GitOrigin-RevId: 110d7aa479ab5680a71de0d3ed22858f62e542eb