Only need to compare the sizes of visible rects
The origins of the visible rects do not matter in these comparisons. The
code only uses visible_data() on the video frames and the width() and
height) of the visible rects.

Also replace a visible_rect local variable with a temporary object.

Change-Id: I5a43b7ab3095e07d9184d35b5e3dfc9b33e19c35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3551881
Reviewed-by: Eugene Zemtsov <eugene@chromium.org>
Commit-Queue: Wan-Teh Chang <wtc@google.com>
Cr-Commit-Position: refs/heads/main@{#985146}
NOKEYCHECK=True
GitOrigin-RevId: c89cfb81b984dcbcf672857424f5aaf08c8f23fb