Add basic browser plumbing for OOP video decoding.
This CL adds the basic browser-process-side plumbing to support
out-of-process video decoding (which includes the LaCrOS use case).

The general approach is to add an optional argument to the
media::mojom::InterfaceFactory::CreateVideoDecoder() method. When the
browser process needs to broker the creation of a new
media::mojom::VideoDecoder (e.g., on behalf of a renderer), we use this
extra argument to pass a pending_remote<StableVideoDecoder>. The GPU
process receives this and (in a separate CL) will use it to act as a
proxy between the renderer and a StableVideoDecoder implementation.

There are two scenarios to consider, Linux/non-LaCrOS and LaCrOS:

- Linux/non-LaCrOS: in order to create a StableVideoDecoder to pass to
the GPU process, the browser process starts a utility process that hosts
a StableVideoDecoderFactory. This utility process is to act as the video
decoder process. This CL requests this process to start, but it
currently won't because we don't have an entry point for it in
content/utility/services.cc yet. That's left for a follow-up CL.

- LaCrOS: the browser is not supposed to start a utility process.
Instead, it's supposed to use the crosapi to ask ash-chrome for a
StableVideoDecoderFactory. Then, ash-chrome can decide whether it wants
to create a new process or use its GPU process to host the
StableVideoDecoderFactory. This LaCrOS path is not implemented in this
CL. That's left as TODOs for pmolinalopez.

This path is guarded behind a disabled-by-default feature flag.

Bug: b:195769334
Test: None
Change-Id: Ic0d2340dad1577c4043f394d5a516c2e424162a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3138304
Reviewed-by: Alex Gough <ajgo@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Wei-Yin Chen <wychen@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#988126}
NOKEYCHECK=True
GitOrigin-RevId: 73a4f15cb3b5acdbd19efb4370d28aaeaac77216