VaapiImageProcessorBackend: Add rotation support test to Create()
This CL addresses an issue where the VaapiImageProcessorBackend could be
configured with an unsupported relative rotation. When the the image
processor actually runs, it will error with a message "VaapiIp doesn't
support rotation". The change adds a check for rotation support to
Create() if the relative rotation is not zero. If the check fails,
ImageProcessorFactory will fall back to libyuv.

Bug: b:228054158
Test: image_processor_test on zork
Change-Id: I0affc98049999e48259a9ffea2d779fe7b5d416b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3569849
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Nathan Hebert <nhebert@chromium.org>
Cr-Commit-Position: refs/heads/main@{#989058}
NOKEYCHECK=True
GitOrigin-RevId: 98a859550cb79af0a5b744d98485260ddf1140f0